fix compiling in 64 bits arch

+++ freedo/Iso.cpp
@@ -24,6 +24,7 @@ Felix Lazarev
 // Iso.cpp: implementation of the CIso class.
 //
 //////////////////////////////////////////////////////////////////////
+#include <stdint.h>
 #include "freedoconfig.h"
 #include <memory.h>
 #include "IsoXBUS.h"
@@ -1337,7 +1338,7 @@ void* _xbplug_MainDevice(int proc, void* data)
 		        isodrive.InitCD();
 		break;
 	case XBP_SET_COMMAND:
-		isodrive.SendCommand((unsigned int)data);
+		isodrive.SendCommand((intptr_t)data);
 		break;
 	case XBP_FIQ:
 		return (void*)isodrive.TestFIQ();
@@ -1346,7 +1347,7 @@ void* _xbplug_MainDevice(int proc, void* data)
 	case XBP_GET_STATUS:
 		return (void*)isodrive.GetStatusFifo();
 	case XBP_SET_POLL:
-		isodrive.SetPoll((unsigned int)data);
+		isodrive.SetPoll((intptr_t)data);
 		break;
 	case XBP_GET_POLL:
 		return (void*)isodrive.GetPoll();
